requirements.txt
langchain
langchain-community
langchain-core
langchain_ollama
pillow
pandas
pydantic